+++ packages/mds-types/policy.ts
@@ -1,4 +1,4 @@
-import { DAY_OF_WEEK, Enum, Timestamp, UUID } from './base'
+import { DAY_OF_WEEK, Enum, Timestamp, UUID } from './utils'
 import { ACCESSIBILITY_OPTION, MODALITY } from './device'
 import { MICRO_MOBILITY_VEHICLE_EVENT, TAXI_VEHICLE_EVENT, TNC_VEHICLE_EVENT } from './event'
 import { MICRO_MOBILITY_VEHICLE_STATE, TAXI_VEHICLE_STATE, TNC_VEHICLE_STATE } from './vehicle/vehicle_states'
